Prevent haxxxing

+++ app/controllers/products_controller.rb
@@ -24,44 +24,64 @@ class ProductsController < ApplicationController
   end
 
   def new
-    render "new.html.erb"
+    if current_user && current_user.admin
+      render "new.html.erb"
+    else
+      redirect_to "/products"
+    end
   end
 
   def create
-    cheese = Product.new(
-      name: params["name"],
-      price: params["price"],
-      image: params["image"],
-      description: params["description"]
-      )
-    cheese.save
-    flash[:success] = "Product added successfully!"
-    redirect_to "/checkout"
+    if current_user && current_user.admin
+      cheese = Product.new(
+        name: params["name"],
+        price: params["price"],
+        image: params["image"],
+        description: params["description"]
+        )
+      cheese.save
+      flash[:success] = "Product added successfully!"
+      redirect_to "/checkout"
+    else
+      redirect_to "/login"
+    end
   end
 
   def edit
-    product_id = params[:id]
-    @product = Product.find_by(id: product_id)
-    render "edit.html.erb"
+    if current_user
+      product_id = params[:id]
+      @product = Product.find_by(id: product_id)
+      render "edit.html.erb"
+    else
+      redirect_to "/products"
+    end
   end
 
   def update
-    product_id = params[:id]
-    product = Product.find_by(id: product_id)
-    product.name = params[:name]
-    product.price = params[:price]
-    product.image = params[:image]
-    product.description = params[:description]
-    product.save
-    flash[:success] = "Product update complete!"
-    redirect_to "/products/#{product.id}"
+    if current_user && current_user.admin
+      product_id = params[:id]
+      product = Product.find_by(id: product_id)
+      product.name = params[:name]
+      product.price = params[:price]
+      product.image = params[:image]
+      product.description = params[:description]
+      product.save
+      flash[:success] = "Product update complete!"
+      redirect_to "/products/#{product.id}"
+    else
+      redirect_to "/login"
+    end
   end
 
   def destroy
-    product_id = params[:id]
-    product = Product.find_by(id: product_id)
-    product.destroy
-    flash[:success] = "Goodbye, out-of-date product"
-    redirect_to "/products"
+    if current_user && current_user.admin
+      product_id = params[:id]
+      product = Product.find_by(id: product_id)
+      product.destroy
+      flash[:success] = "Goodbye, out-of-date product"
+      redirect_to "/products"
+    else
+      redirect_to "/login"
+    end
   end
 end
